Added controller

+++ WarOfRaces/src/edu/QATestLab/WarOfRaces/controller/Teams.java
@@ -35,41 +35,41 @@ public class Teams {
 
 	private void makeTeams() {
 		
-		alElvesTeam.add(new Mage((byte)100, "Mage", (byte)0, (byte)10));
-		alElvesTeam.add(new Archer((byte)100, "Archer1", (byte)7, (byte)3));
-		alElvesTeam.add(new Archer((byte)100, "Archer2", (byte)7, (byte)3));
-		alElvesTeam.add(new Archer((byte)100, "Archer3", (byte)7, (byte)3));
-		alElvesTeam.add(new Warrior((byte)100, "Warrior1", (byte)0, (byte)15));
-		alElvesTeam.add(new Warrior((byte)100, "Warrior2", (byte)0, (byte)15));
-		alElvesTeam.add(new Warrior((byte)100, "Warrior3", (byte)0, (byte)15));
-		alElvesTeam.add(new Warrior((byte)100, "Warrior4", (byte)0, (byte)15));
+		alElvesTeam.add(new Mage(100, "Mage", 0, 10));
+		alElvesTeam.add(new Archer(100, "Archer1", 7, 3));
+		alElvesTeam.add(new Archer(100, "Archer2", 7, 3));
+		alElvesTeam.add(new Archer(100, "Archer3", 7, 3));
+		alElvesTeam.add(new Warrior(100, "Warrior1", 0, 15));
+		alElvesTeam.add(new Warrior(100, "Warrior2", 0, 15));
+		alElvesTeam.add(new Warrior(100, "Warrior3", 0, 15));
+		alElvesTeam.add(new Warrior(100, "Warrior4", 0, 15));
 		
-		alHumanTeam.add(new MageHuman((byte)100, "MageHuman", (byte)0, (byte)4));
-		alHumanTeam.add(new CrossbowArcher((byte)100, "CrossbowArcher1", (byte)5, (byte)3));
-		alHumanTeam.add(new CrossbowArcher((byte)100, "CrossbowArcher2", (byte)5, (byte)3));
-		alHumanTeam.add(new CrossbowArcher((byte)100, "CrossbowArcher3", (byte)5, (byte)3));
-		alHumanTeam.add(new WarriorHuman((byte)100, "WarriorHuman1", (byte)0, (byte)18));
-		alHumanTeam.add(new WarriorHuman((byte)100, "WarriorHuman2", (byte)0, (byte)18));
-		alHumanTeam.add(new WarriorHuman((byte)100, "WarriorHuman3", (byte)0, (byte)18));
-		alHumanTeam.add(new WarriorHuman((byte)100, "WarriorHuman4", (byte)0, (byte)18));
+		alHumanTeam.add(new MageHuman(100, "MageHuman", 0, 4));
+		alHumanTeam.add(new CrossbowArcher(100, "CrossbowArcher1", 5, 3));
+		alHumanTeam.add(new CrossbowArcher(100, "CrossbowArcher2", 5, 3));
+		alHumanTeam.add(new CrossbowArcher(100, "CrossbowArcher3", 5, 3));
+		alHumanTeam.add(new WarriorHuman(100, "WarriorHuman1", 0, 18));
+		alHumanTeam.add(new WarriorHuman(100, "WarriorHuman2", 0, 18));
+		alHumanTeam.add(new WarriorHuman(100, "WarriorHuman3", 0, 18));
+		alHumanTeam.add(new WarriorHuman(100, "WarriorHuman4", 0, 18));
 		
-		alOrcsTeam.add(new Shaman((byte)100, "Shaman", (byte)0, (byte)0));
-		alOrcsTeam.add(new ArcherOrc((byte)100, "ArcherOrc1", (byte)3, (byte)2));
-		alOrcsTeam.add(new ArcherOrc((byte)100, "ArcherOrc2", (byte)3, (byte)2));
-		alOrcsTeam.add(new ArcherOrc((byte)100, "ArcherOrc3", (byte)3, (byte)2));
-		alOrcsTeam.add(new Goblin((byte)100, "Goblin1", (byte)0, (byte)20));
-		alOrcsTeam.add(new Goblin((byte)100, "Goblin2", (byte)0, (byte)20));
-		alOrcsTeam.add(new Goblin((byte)100, "Goblin3", (byte)0, (byte)20));
-		alOrcsTeam.add(new Goblin((byte)100, "Goblin4", (byte)0, (byte)20));
+		alOrcsTeam.add(new Shaman(100, "Shaman", 0, 0));
+		alOrcsTeam.add(new ArcherOrc(100, "ArcherOrc1", 3, 2));
+		alOrcsTeam.add(new ArcherOrc(100, "ArcherOrc2", 3, 2));
+		alOrcsTeam.add(new ArcherOrc(100, "ArcherOrc3", 3, 2));
+		alOrcsTeam.add(new Goblin(100, "Goblin1", 0, 20));
+		alOrcsTeam.add(new Goblin(100, "Goblin2", 0, 20));
+		alOrcsTeam.add(new Goblin(100, "Goblin3", 0, 20));
+		alOrcsTeam.add(new Goblin(100, "Goblin4", 0, 20));
 		
-		alUndeadsTeam.add(new Necromancer((byte)100, "Necromancer", (byte)0, (byte)5));
-		alUndeadsTeam.add(new Hunter((byte)100, "Hunter1", (byte)4, (byte)2));
-		alUndeadsTeam.add(new Hunter((byte)100, "Hunter2", (byte)4, (byte)2));
-		alUndeadsTeam.add(new Hunter((byte)100, "Hunter3", (byte)4, (byte)2));
-		alUndeadsTeam.add(new Zombie((byte)100, "Zombie1", (byte)0, (byte)18));
-		alUndeadsTeam.add(new Zombie((byte)100, "Zombie2", (byte)0, (byte)18));
-		alUndeadsTeam.add(new Zombie((byte)100, "Zombie3", (byte)0, (byte)18));
-		alUndeadsTeam.add(new Zombie((byte)100, "Zombie4", (byte)0, (byte)18));
+		alUndeadsTeam.add(new Necromancer(100, "Necromancer", 0, 5));
+		alUndeadsTeam.add(new Hunter(100, "Hunter1", 4, 2));
+		alUndeadsTeam.add(new Hunter(100, "Hunter2", 4, 2));
+		alUndeadsTeam.add(new Hunter(100, "Hunter3", 4, 2));
+		alUndeadsTeam.add(new Zombie(100, "Zombie1", 0, 18));
+		alUndeadsTeam.add(new Zombie(100, "Zombie2", 0, 18));
+		alUndeadsTeam.add(new Zombie(100, "Zombie3", 0, 18));
+		alUndeadsTeam.add(new Zombie(100, "Zombie4", 0, 18));
 	}
 	
 	public ArrayList<Unit> addTeamOpponents1(){
