+++ WarOfRaces/src/edu/QATestLab/WarOfRaces/model/Elves/Archer.java
@@ -2,28 +2,28 @@ package edu.QATestLab.WarOfRaces.model.Elves;
 
 public class Archer extends RaceOfElves{
 	
-	private byte damageFirst;
-	private byte damageSecond;
+	private float damageFirst;
+	private float damageSecond;
 	
 	
 
-	public byte getDamageFirst() {
+	public float getDamageFirst() {
 		return damageFirst;
 	}
 
-	public void setDamageFirst(byte damageFirst) {
+	public void setDamageFirst(float damageFirst) {
 		this.damageFirst = damageFirst;
 	}
 
-	public byte getDamageSecond() {
+	public float getDamageSecond() {
 		return damageSecond;
 	}
 
-	public void setDamageSecond(byte damageSecond) {
+	public void setDamageSecond(float damageSecond) {
 		this.damageSecond = damageSecond;
 	}
 
-	public Archer(byte hp, String name, byte damageFirst, byte damageSecond) {
+	public Archer(float hp, String name, float damageFirst, float damageSecond) {
 		super(hp, name, damageSecond, damageSecond);
 		this.damageFirst = damageFirst;
 		this.damageSecond = damageSecond;
@@ -31,7 +31,7 @@ public class Archer extends RaceOfElves{
 
 	@Override
 	public void attackFirst(Unit u) {
-		u.setHp((byte) (u.getHp() - damageFirst));
+		u.setHp((u.getHp() - damageFirst));
 	}
 
 	@Override
