+++ WarOfRaces/src/edu/QATestLab/WarOfRaces/model/Elves/Mage.java
@@ -3,32 +3,32 @@ package edu.QATestLab.WarOfRaces.model.Elves;
 
 public class Mage extends RaceOfElves {
 
-	private byte damageSecond;
+	private float damageSecond;
 	
 	
-	public byte getDamageSecond() {
+	public float getDamageSecond() {
 		return damageSecond;
 	}
 
-	public void setDamageSecond(byte damageSecond) {
+	public void setDamageSecond(float damageSecond) {
 		this.damageSecond = damageSecond;
 	}
 
-	public Mage(byte hp, String name, byte damageFirst, byte damageSecond) {
+	public Mage(float hp, String name, float damageFirst, float damageSecond) {
 		super(hp, name, damageSecond, damageSecond);
 		this.damageSecond = damageSecond;
 	}
 
 	@Override
 	public void attackFirst(Unit u) {
-		u.setDamageFirst((byte) (u.getDamageFirst() * 1.5));
-		u.setDamageSecond((byte) (getDamageSecond() * 1.5));
+		u.setDamageFirst((float) (u.getDamageFirst() * 1.5));
+		u.setDamageSecond((float) (getDamageSecond() * 1.5));
 		System.out.println("____________");
 	}
 
 	@Override
 	public void attackSecond(Unit u) {
-		u.setHp((byte) (u.getHp() - damageSecond));
+		u.setHp((float) (u.getHp() - damageSecond));
 	}
