+++ WarOfRaces/src/edu/QATestLab/WarOfRaces/model/Elves/RaceOfElves.java
@@ -4,9 +4,11 @@ package edu.QATestLab.WarOfRaces.model.Elves;
 
 public class RaceOfElves extends Unit{
 
-	public RaceOfElves(byte hp, String name, byte damageFirst, byte damageSecond) {
+	public RaceOfElves(float hp, String name, float damageFirst, float damageSecond) {
 		super(hp, name, damageSecond, damageSecond);
 	}
+
+	
 	
 
 }
