+++ WarOfRaces/src/edu/QATestLab/WarOfRaces/model/Human/CrossbowArcher.java
@@ -5,28 +5,28 @@ import edu.QATestLab.WarOfRaces.model.Elves.Unit;
 
 public class CrossbowArcher extends RaceOfHuman{
 	
-	private byte damageFirst;
-	private byte damageSecond;
+	private float damageFirst;
+	private float damageSecond;
 	
 	
 	
-	public byte getDamageFirst() {
+	public float getDamageFirst() {
 		return damageFirst;
 	}
 
-	public void setDamageFirst(byte damageFirst) {
+	public void setDamageFirst(float damageFirst) {
 		this.damageFirst = damageFirst;
 	}
 
-	public byte getDamageSecond() {
+	public float getDamageSecond() {
 		return damageSecond;
 	}
 
-	public void setDamageSecond(byte damageSecond) {
+	public void setDamageSecond(float damageSecond) {
 		this.damageSecond = damageSecond;
 	}
 
-	public CrossbowArcher(byte hp, String name, byte damageFirst, byte damageSecond) {
+	public CrossbowArcher(float hp, String name, float damageFirst, float damageSecond) {
 		super(hp, name, damageSecond, damageSecond);
 		this.damageFirst = damageFirst;
 		this.damageSecond = damageSecond;
@@ -34,12 +34,12 @@ public class CrossbowArcher extends RaceOfHuman{
 
 	@Override
 	public void attackFirst(Unit u) {
-		u.setHp((byte) (u.getHp() - damageFirst));
+		u.setHp((float) (u.getHp() - damageFirst));
 	}
 
 	@Override
 	public void attackSecond(Unit u) {
-		u.setHp((byte) (u.getHp() - damageSecond));
+		u.setHp((float) (u.getHp() - damageSecond));
 	}
