+++ WarOfRaces/src/edu/QATestLab/WarOfRaces/model/Orcs/Goblin.java
@@ -4,26 +4,26 @@ import edu.QATestLab.WarOfRaces.model.Elves.Unit;
 
 public class Goblin extends RaceOfOrcs {
 	
-	private byte damageSecond;
+	private float damageSecond;
 	
 	
 	
-	public byte getDamageSecond() {
+	public float getDamageSecond() {
 		return damageSecond;
 	}
 
-	public void setDamageSecond(byte damageSecond) {
+	public void setDamageSecond(float damageSecond) {
 		this.damageSecond = damageSecond;
 	}
 
-	public Goblin(byte hp, String name, byte damageFirst, byte damageSecond) {
+	public Goblin(float hp, String name, float damageFirst, float damageSecond) {
 		super(hp, name, damageSecond, damageSecond);
 		this.damageSecond = damageSecond;
 	}
 
 	@Override
 	public void attackFirst(Unit u) {
-		u.setHp((byte) (u.getHp() - damageSecond));
+		u.setHp((float) (u.getHp() - damageSecond));
 	}
 
 	@Override
