+++ WarOfRaces/src/edu/QATestLab/WarOfRaces/model/Undead/Necromancer.java
@@ -4,22 +4,22 @@ import edu.QATestLab.WarOfRaces.model.Elves.Unit;
 
 public class Necromancer extends RaceOfUndead {
 	
-	private byte damageSecond;
+	private float damageSecond;
 	
-	public Necromancer(byte hp, String name, byte damageFirst, byte damageSecond) {
+	public Necromancer(float hp, String name, float damageFirst, float damageSecond) {
 		super(hp, name, damageSecond, damageSecond);
 		this.damageSecond = damageSecond;
 	}
 
 	@Override
 	public void attackFirst(Unit u) {
-		u.setDamageFirst((byte) (u.getDamageFirst() / 2));
-		u.setDamageSecond((byte) (u.getDamageSecond() / 2));
+		u.setDamageFirst((float) (u.getDamageFirst() / 2));
+		u.setDamageSecond((float) (u.getDamageSecond() / 2));
 	}
 
 	@Override
 	public void attackSecond(Unit u) {
-		u.setHp((byte) (u.getHp() - damageSecond));
+		u.setHp((float) (u.getHp() - damageSecond));
 	}
